+++ CHANGELOG.md
@@ -7,9 +7,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- Switched to using `cluster-standup-teardown` to handle cluster creation and deletion logic
+- Removed cluster and default-apps values in favour of defaults from `cluster-standup-teardown`
+
 ## [1.38.0] - 2024-04-25
 
 ### Changed
+
 - CAPZ: Change Helm values structure where root-level properties are moved to `.Values.global`.
 
 ## [1.37.0] - 2024-04-19